+++ src/editor-model/delete.ts
@@ -75,7 +75,7 @@ import { contentWillChange } from './listeners';
 // }
 
 /**
- * Handle special cases when deleting an atom as per the table below
+ * Handle special cases when deleting an atom as per the table below:
  * - deleting an empty numerator: demote fraction
  * - forward-deleting a square root: demote it
  * - delete last atom inside a square root: delete the square root
@@ -83,7 +83,7 @@ import { contentWillChange } from './listeners';
  * - etc...
  *
  *
- * @param branch: if deleting inside an atom, the branch being delete
+ * @param branch: if deleting inside an atom, the branch being deleted
  * (always the first or last atom of the branch). If undefined, the atom
  * itself is about to be deleted.
  *
@@ -96,6 +96,26 @@ function onDelete(
   branch?: Branch
 ): boolean {
   const parent = atom.parent!;
+
+  if (parent.type === 'genfrac') {
+    let pos = model.offsetOf(atom.leftSibling);
+    parent.removeChild(atom);
+
+    if (parent.hasEmptyBranch('above') && parent.hasEmptyBranch('below')) {
+      // The last numerator or denominator of a fraction has been deleted:
+      // delete the fraction
+
+      pos = model.offsetOf(parent.leftSibling);
+      parent.parent!.removeChild(parent);
+      model.announce('delete', undefined, [parent]);
+      model.position = pos;
+      return true;
+    }
+    model.announce('delete', undefined, [atom]);
+    model.position = pos;
+    return true;
+  }
+
   if (atom instanceof LeftRightAtom) {
     //
     // 'leftright': \left\right