+++ src/editor-mathfield/mathfield-private.ts
@@ -931,9 +931,9 @@ export class MathfieldPrivate implements GlobalContext, Mathfield {
     return ce.box(ce.parse(this.model.getValue()));
   }
 
-  loadSound(
+  async loadSound(
     sound: 'keypress' | 'spacebar' | 'delete' | 'plonk' | 'return'
-  ): void {
+  ): Promise<void> {
     //  Clear out the cached audio buffer
     delete this.audioBuffers[sound];
 
@@ -977,18 +977,18 @@ export class MathfieldPrivate implements GlobalContext, Mathfield {
       return;
 
     // Fetch the audio buffer
-    fetch(resolveUrl(soundsDirectory + '/' + soundFile))
-      .then((response) => response.arrayBuffer())
-      .then((arrayBuffer) => this.audioContext.decodeAudioData(arrayBuffer))
-      .then((audioBuffer) => {
-        this.audioBuffers[sound] = audioBuffer;
-      });
+    const response = await fetch(
+      await resolveUrl(soundsDirectory + '/' + soundFile)
+    );
+    const arrayBuffer = await response.arrayBuffer();
+    const audioBuffer = await this.audioContext.decodeAudioData(arrayBuffer);
+    this.audioBuffers[sound] = audioBuffer;
   }
 
-  playSound(
+  async playSound(
     name: 'keypress' | 'spacebar' | 'delete' | 'plonk' | 'return'
-  ): void {
-    if (!this.audioBuffers[name]) this.loadSound(name);
+  ): Promise<void> {
+    if (!this.audioBuffers[name]) await this.loadSound(name);
     if (!this.audioBuffers[name]) return;
 
     // A sound source can't be played twice, so creeate a new one
@@ -1096,7 +1096,7 @@ export class MathfieldPrivate implements GlobalContext, Mathfield {
         if (this.options.keypressVibration && canVibrate())
           navigator.vibrate(HAPTIC_FEEDBACK_DURATION);
 
-        void this.playSound('keypress');
+        this.playSound('keypress');
       }
 
       if (options.scrollIntoView) this.scrollIntoView();