+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - **#1726** Deleting the last element of a fraction also deletes the fraction
 - **#1764** The MathML serialization for superscripts and subscripts was
   invalid.
+- **#1790** Annotations from the `\enclose` command could not be displayed in
+  some cases if the `z-index` of the expression they decorated had certain
+  values.
 
 ## 0.86.0 (2022-12-02)